refactor: extract supertest types
