+++ src/testing/helpers/e2e.ts
@@ -6,7 +6,6 @@ import { Test } from '@nestjs/testing';
 import { MongoMemoryReplSet } from 'mongodb-memory-server';
 import { fromAsyncThrowable } from 'neverthrow';
 import request from 'supertest';
-import type SupertestAgent from 'supertest/lib/agent.js';
 import { afterAll, beforeAll, beforeEach, describe, vi } from 'vitest';
 import type { SuiteAPI } from 'vitest';
 
@@ -15,45 +14,12 @@ import { loadAppContainer, loadUserConfig } from '../../meta/load.js';
 import { MONGO_CONNECTION_TOKEN } from '../../modules/prisma/prisma.config.js';
 
 import type { MongoConnection } from '../../modules/prisma/connection.factory.js';
-
-interface TestResponse {
-  [key: string]: any;
-  body: any;
-  headers: {
-    [key: string]: string;
-  };
-  ok: boolean;
-  status: number;
-  text: string;
-  type: string;
-}
-
-interface TestRequest extends PromiseLike<TestResponse> {
-  [key: string]: any;
-  accept(type: string): this;
-  method: string;
-  send(data?: object | string): this;
-  set(field: string, val: string): this;
-  url: string;
-}
-
-interface TestAgentMethods {
-  delete: (url: string) => TestRequest;
-  get: (url: string) => TestRequest;
-  patch: (url: string) => TestRequest;
-  post: (url: string) => TestRequest;
-  put: (url: string) => TestRequest;
-}
-
-// this is so we don't have to include the garbage supertest types in production
-type TestAgent = SupertestAgent extends TestAgentMethods
-  ? TestAgentMethods & {
-      setAccessToken: (token: string) => void;
-    }
-  : never;
+import type { TestAgent } from './types.js';
 
 export type EndToEndContext = {
-  api: TestAgent;
+  api: TestAgent<{
+    setAccessToken: (token: string) => void;
+  }>;
 };
 
 export function e2e(fn: (describe: SuiteAPI<EndToEndContext>) => void): void {
